pyproject.toml
[tool.poetry]
name = "habitsync"
version = "0.1.0"
description = ""
authors = ["dima <dmitrij-bezgubov@yandex.ru>"]
readme = "README.md"
packages = [{ include = "habitsync" }]  # Указывает, что папка habitsync - это пакет

[tool.poetry.dependencies]
python = "^3.12"
django = "^5.2"
psycopg2 = "^2.9.10"
djangorestframework = "^3.16.0"
djangorestframework-simplejwt = ">=5.5.0,<6.0.0"
django-cors-headers = ">=4.7.0,<5.0.0"
drf-yasg = ">=1.21.10,<2.0.0"
pillow = ">=11.2.1,<12.0.0"
dotenv = ">=0.9.9,<0.10.0"
requests = ">=2.32.3,<3.0.0"
celery = ">=5.5.1,<6.0.0"
django-filter = ">=25.1,<26.0"
django-celery-beat = { version = ">=2.8.0,<3.0.0", python = ">=3.12,<4.0" }
ipython = ">=9.1.0,<10.0.0"
redis = ">=5.2.1,<6.0.0"
eventlet = ">=0.39.1,<0.40.0"
python-dotenv = ">=1.1.0,<2.0.0"
pytest = "^8.3.5"
pytest-django = "^4.11.1"
pytest-cov = "^6.1.1"



[tool.poetry.group.dev.dependencies]
pytest = "^8.3.5"
pytest-django = "^4.11.1"
pytest-cov = "^6.1.1"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"